init role check

+++ src/employee/employee.controller.ts
@@ -6,22 +6,29 @@ import {
   Patch,
   Param,
   Delete,
+  Request,
   UseGuards,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { EmployeeService } from './employee.service';
 import { CreateEmployeeDto } from './dto/create-employee.dto';
 import { UpdateEmployeeDto } from './dto/update-employee.dto';
 import { JwtAuthGuard } from '..//user/auth/jwt-auth.guard';
+import { UserService } from '../user/user.service';
 
 @Controller('api/employee')
 export class EmployeeController {
-  constructor(private readonly employeeService: EmployeeService) {}
+  constructor(
+    private readonly employeeService: EmployeeService,
+    private readonly userService: UserService,
+  ) {}
 
   @UseGuards(JwtAuthGuard)
   @Post()
-  async create(@Body() createEmployeeDto: CreateEmployeeDto) {
+  async create(@Body() createEmployeeDto: CreateEmployeeDto, @Request() req) {
+    await this.checkPermission(req);
     return this.employeeService.create(createEmployeeDto);
-  }
+  }  
 
   @Get()
   @UseGuards(JwtAuthGuard)
@@ -40,13 +47,23 @@ export class EmployeeController {
   async update(
     @Param('id') id: string,
     @Body() updateEmployeeDto: UpdateEmployeeDto,
+    @Request() req,
   ) {
+    await this.checkPermission(req);
     return this.employeeService.update(id, updateEmployeeDto);
   }
 
   @Delete(':id')
   @UseGuards(JwtAuthGuard)
-  async remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string, @Request() req) {
+    await this.checkPermission(req);
     return this.employeeService.remove(id);
+  }
+
+  private async checkPermission(req: any) {
+    const user = await this.userService.findById(req.user.userId);
+    if (user.role != 'admin') {
+      throw new UnauthorizedException("You don't have enough permission");
+    }
   }
 }
